:art: Format the code using black

+++ liteasr/nets/gumbel_vector_quantizer.py
@@ -9,7 +9,6 @@ import torch.nn.functional as F
 
 
 class GumbelVectorQuantizer(nn.Module):
-
     def __init__(
         self,
         dim: int,
@@ -45,9 +44,7 @@ class GumbelVectorQuantizer(nn.Module):
         var_dim = vq_dim // groups
         num_groups = groups if not combine_groups else 1
 
-        self.vars = nn.Parameter(
-            torch.FloatTensor(1, num_groups * num_vars, var_dim)
-        )
+        self.vars = nn.Parameter(torch.FloatTensor(1, num_groups * num_vars, var_dim))
         nn.init.uniform_(self.vars)
 
         self.weight_proj = nn.Linear(self.input_dim, groups * num_vars)
@@ -56,6 +53,7 @@ class GumbelVectorQuantizer(nn.Module):
 
         if isinstance(temp, str):
             import ast
+
             temp = ast.literal_eval(temp)
         assert len(temp) == 3, f"{temp}, {len(temp)}"
 
@@ -81,8 +79,9 @@ class GumbelVectorQuantizer(nn.Module):
         # compute codebook index
         _, k = x.max(-1)  # (b * t * g)
         hard_x = (
-            x.new_zeros(*x.shape).scatter_(-1, k.view(-1, 1),
-                                           1.0).view(b * t, self.groups, -1)
+            x.new_zeros(*x.shape)
+            .scatter_(-1, k.view(-1, 1), 1.0)
+            .view(b * t, self.groups, -1)
         )  # (b * t, g, nv)
 
         avg_x = torch.softmax(
@@ -91,9 +90,9 @@ class GumbelVectorQuantizer(nn.Module):
         avg_probs = torch.mean(avg_x, dim=0)  # (g, nv)
 
         if self.training:
-            x = F.gumbel_softmax(
-                x.float(), tau=self.curr_temp, hard=True
-            ).type_as(x)  # (b * t * g, nv)
+            x = F.gumbel_softmax(x.float(), tau=self.curr_temp, hard=True).type_as(
+                x
+            )  # (b * t * g, nv)
         else:
             x = hard_x
 
@@ -105,9 +104,7 @@ class GumbelVectorQuantizer(nn.Module):
 
         # concatenate
         x = x.unsqueeze(-1) * vars  # (b * t, g * nv, vq_dim // g)
-        x = x.view(
-            b * t, self.groups, self.num_vars, -1
-        )  # (b * t, g, nv, vq_dim // g)
+        x = x.view(b * t, self.groups, self.num_vars, -1)  # (b * t, g, nv, vq_dim // g)
         x = x.sum(-2)  # (b * t, g, vq_dim // g)
         x = x.view(b, t, -1)  # (b, t, vq_dim)
